Fixed Full Throttle, update to 1.3.7 release
Turns out that Full Throttle Remastered uses 8 bytes for the offset so
for files at the end of the bundle the offset was overflowing and became
wrong. So changed offset to be a 64 bit value.
OGV filetype is actually ogg video after all - previous wrong offsets
meant that it was grabbing data inside FSB's making me think they were
mp3 data.
Updated readme for V1.3.7